backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
anthropic==0.40.0
PyGithub==2.4.0
pydantic==2.9.2
networkx==3.3
python-multipart==0.0.12
fastembed==0.3.6
numpy>=1.26.0
